+++ src/DDTrace/SpanContext.php
@@ -26,7 +26,7 @@ final class SpanContext extends SpanContextData
     {
         $instance = new self(
             $parentContext->getTraceId(),
-            ID::generate(),
+            dd_trace_generate_id(),
             $parentContext->getSpanId(),
             $parentContext->getAllBaggageItems(),
             false
@@ -38,7 +38,7 @@ final class SpanContext extends SpanContextData
 
     public static function createAsRoot(array $baggageItems = [])
     {
-        $nextId = ID::generate();
+        $nextId = dd_trace_generate_id();
 
         return new self(
             $nextId,